+++ src/index.ts
@@ -1,4 +1,4 @@
-export { LoggerModule } from './LoggerModule';
+export { LoggerModule } from './logger.module';
 export { Params, LoggerModuleAsyncParams } from './params';
 export { Logger, PinoLogger } from './services';
 export { InjectPinoLogger, getLoggerToken } from './InjectPinoLogger';
